Migrate Query.fieldStats
